fir_email/requirements_smime.txt
django-djembe==0.2.0